[Kernel][Defaults] Remove redundant `FileIO.readFiles`; instead use `FileIO.newInputFile` for reads
